+++ lib/bademagic_module/bluetooth/scanstate.dart
@@ -3,7 +3,7 @@ import 'package:badgemagic/bademagic_module/bluetooth/connectstate.dart';
 import 'package:flutter_blue_plus/flutter_blue_plus.dart';
 import 'package:logger/logger.dart';
 import 'ble_state_interface.dart';
-import 'toast.dart';
+import 'bletoast.dart';
 
 class ScanState implements BleState {
   ScanResult? foundDevice;
